fix: 스와이프 알림 팝업 위치 변경
